feat(authentication): persisting authed state in the front end
