+++ components/AuthCheck/index.ts
@@ -1,0 +1,1 @@
+export { AuthCheck } from "./AuthCheck";
