+++ context/AuthContext.tsx
@@ -1,9 +1,10 @@
 "use client"
 
 // REACT
-import { createContext, useState } from "react";
+import { createContext, useState, useEffect } from "react";
 // FIREBASE
 import firebase from "firebase/app";
+import { auth } from "@/firebase/client/config";
 
 type AuthContextType = {
   user: firebase.User | null;
@@ -15,6 +16,15 @@ const AuthContext = createContext<AuthContextType | null>(null);
 const AuthContextProvider = ({ children }: { children: React.ReactNode }) => {
   // STATE && VARIABLES
   const [user, setUser] = useState<firebase.User | null>(null);
+
+  // USE EFFECTS
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((user) => {
+      user ? setUser(user) : setUser(null);
+    })
+
+    return () => unsubscribe();
+  }, [user]);
 
   return (
     <AuthContext.Provider value={{ user, setUser }}>
